feat(sales): modela venda com domínio rico e histórico de faturamento
- Cria entidade de domínio Sale com factories create/restore e regras de negócio
- Adiciona snapshot de pessoa e plano na venda
- Substitui flag de billing por BillingHistory como Value Object
- Cria StatusPaymentEnum e histórico de pagamentos
- Ajusta entidades JPA (SaleJpa e BillingHistoryJpa) e relacionamento OneToMany
- Implementa mapeamento completo no SaleRepositoryAdapter
- Separa configuração de cobrança (DTO) de histórico de pagamentos
- Ajusta fluxo de criação da venda no service

+++ backend/sales-service/src/main/java/com/salesflow/sales_service/infrastructure/gateway/dto/PlanDto.java
@@ -1,10 +1,12 @@
 package com.salesflow.sales_service.infrastructure.gateway.dto;
 
+import java.math.BigDecimal;
+
 public record PlanDto(
         String planId,
         String name,
         String type,
-        double monthlyPrice,
+        BigDecimal monthlyPrice,
         boolean active,
         String description
 ) {}
